코드 리팩토링 및 유틸클래스 분리
